requirements.txt
datasets==2.19.0
fasttext_wheel==0.9.2
